Add fn to fetch offset of input coin data
The interpreter will need to access the serialized transaction input
data for coin type when executing the predicate verification.

For this, we need an encapsulated method that will, provided an input
index, return its relative offset in the serialized bytes.

Besides this change, the enum representation was refactored to reflect
private representations of enumerated identifiers. The ID type was also
split into HASH and ADDRESS.

ContractAddress type was created to differentiate between regular
addresses and contract addresses.