Add `prepare_sign` method to transaction
The transaction to be signed have some fields zeroed as in the specs.

This commit facilitates the bytes manipulation of a serialized
transaction so they can be zeroed with no hardcoded positions.

With the `prepare_sign` function, it is trivial to add `id` considering
some generic hash function - currently defined as Sha256.