Add I/O/W offset functions for transaction
The transaction must provide a set of functions to its users that will
allow the offset of the inputs, outputs and witnesses to be exposed.

These offsets are expected to be used in the interpreter execution. For
that case, the transaction is serialized in-memory, and the execution
should use the offsets provided by these functions.