Add Metadata to optimize tx recurrent functions
A metadata structure will optimize some functions such as `{input,
output, witness}_offset` and `id`.

When `prepare_metadata` is called, it will be stored internally in the
transaction.

The calls to the optimized functions will first query the metadata. If
metadata is not present, then the result will be computed.