Add Metadata to optimize tx recurrent functions (#24)
A metadata structure will optimize some functions such as `{input,
output, witness}_offset` and `id`.

When `prepare_metadata` is called, it will be stored internally in the
transaction.

The calls to the optimized functions will first query the metadata. If
metadata is not present, then the result will be computed.
The transaction is not expected to be mutated in runtime after
deserialization.

The metadata optimization assumes that, once precomputed, it won't need
to recompute. For that, the mutable functions are taken out of the scope of
the transaction so the precomputed metadata won't ever need to be
refreshed.