Add Input::coin_predicate[_data]_offset
This function is needed to calculate predicate/data offsets before the
transaction is created.

This offset can be trivially calculated after the transaction is
created, but this is not desired considering the transaction is not
expected to be mutable.

This function will replace some eventual new opcodes that will evaluate
the transaction at runtime and will set a register with the desired
offset. The following sequence should be enough to set the offset:

```
XIS(0x10, i)
ADDI(0x10, 0x10, Input::coin_predicate_offset)
```