Split `Input::coin` into predicate and signed
The input coin defined in the specifications is an aggregated type
between two logical variants: signed and predicate.

When the input is expected to behave as coin, the signature validation
should seek its witness counterpart and check the signature with its
fields as arguments = except for predicate and its data.

When the input is expected to behave as predicate, the validation will
check if the owner is the hash of the predicate. The witness id is
entirely ignored.

It is canonical to Rust to split these two different logical structures
into different enum variants - even if we are to maintain the fuel specs
canonical serialization.

Closes #110