Add bytecode length to tx
A signed transaction must contain a bytecode length so it won't be
malleable.

Prior to this commit, it checked dynamically; but then the witnesses are
cleared before the tx id computation, so the length is lost.

This commit introduces the attribute that will concretely store the
bytecode len so it will be properly signed and checked.