Add from/to str impls for Transaction
This is required to parse transaction from logs of the VM so this
process can be automated.

Its naturally possible to also leak transaction internals to consumers
of this library so they can serialize//deserialize the transaction
themselves, but its preferable to have a canonical implementation that
can be consumed downstream.